Version 1.13 Files
Version 1.13 fixed a bug in stepwise that arose when StatsModels changed Logit and MNLogit.  They no longer provide the same output from results.pvalues.  Logit still provides a one dimensional pandas like vector.  MNLogit sends back a two dimensional vector.  Indexing is different.  In addition, it's apparent that MNLogit needs work with its convergence routine.  It's failing for simple data.

Changed all of stepwise to  use Logit.  This is fine for binary targets but may be a problem for nominal.